Implemented Activity and StudyingActivity
Activity is an abstract class that serves as the template for all the activites
in the overworld level. It contains most of the variables and the concrete and
abstract methods for implmenting classes. Activity was originally named as
ActivityManager.

StudyingActivity implements Activity. It currently works and performs the
function indicated in the GDD.

Furthermore, a debug function has been added to ClockManager for testing purposes.
It will not deleted in case it is needed again.